+++ src/modules/sqlite_read_mod.f90
@@ -1159,7 +1159,7 @@ contains
   end function getObsFileName
 
 
-  subroutine sqlr_writeAllSqlDiagFiles( obsdat, sfFileName, onlyAssimObs )
+  subroutine sqlr_writeAllSqlDiagFiles( obsdat, sfFileName, onlyAssimObs, addFSOdiag )
     !
     ! :Purpose: To prepare the writing of obsSpaceData content into SQLite format files
     !  
@@ -1169,6 +1169,7 @@ contains
     type(struct_obs)       :: obsdat       ! obsSpaceData object
     character(len=*)       :: sfFileName   ! fileName acronym used for surface obs file
     logical                :: onlyAssimObs ! only write assimilated obs
+    logical                :: addFSOdiag   ! only write FSO
 
     ! locals:
     integer                :: familyIndex, codeTypeIndex, fileIndex
@@ -1222,7 +1223,7 @@ contains
 
           write(*,*) 'tovsCodeTypeListSize = ', tovsCodeTypeListSize
           write(*,*) 'tovsCodeTypeList = ', tovsCodeTypeList(1:tovsCodeTypeListSize) 
-          call sqlr_writeSqlDiagFile(obsdat, 'TO', onlyAssimObs, &
+          call sqlr_writeSqlDiagFile(obsdat, 'TO', onlyAssimObs, addFSOdiag, &
                                      tovsFileNameList(fileIndex), &
                                      tovsCodeTypeList(1:tovsCodeTypeListSize)) 
         end do
@@ -1231,7 +1232,7 @@ contains
 
         fileName = getObsFileName(obsFamilyList(familyIndex), sfFileName_opt=sfFileName)
         call sqlr_writeSqlDiagFile(obsdat, obsFamilyList(familyIndex),  &
-                                   onlyAssimObs, fileName) 
+                                   onlyAssimObs, addFSOdiag, fileName) 
 
       end if   
       
@@ -1337,7 +1338,7 @@ contains
   end subroutine getObsFamilyListMpiGlobal
 
 
-  subroutine sqlr_writeSqlDiagFile( obsdat, obsFamily, onlyAssimObs, &
+  subroutine sqlr_writeSqlDiagFile( obsdat, obsFamily, onlyAssimObs, addFSOdiag, &
                                     instrumentFileName, codeTypeList_opt )
     !
     ! :Purpose: To write the obsSpaceData content into SQLite format files
@@ -1348,6 +1349,7 @@ contains
     type(struct_obs)  :: obsdat
     character(len=*)  :: obsFamily
     logical           :: onlyAssimObs
+    logical           :: addFSOdiag
     character(len=*)  :: instrumentFileName
     integer, optional :: codeTypeList_opt(:)
 
@@ -1357,7 +1359,7 @@ contains
     type(fSQL_STATUS)      :: stat                 ! type for error status
     integer                :: obsVarno, obsFlag, ASS, vertCoordType, codeType, date, time, idObs, idData 
     real                   :: obsValue, OMA, OMP, OER, FGE, PPP, lon, lat, altitude
-    real                   :: ensInnovStdDev, ensObsErrStdDev, zhad
+    real                   :: ensInnovStdDev, ensObsErrStdDev, zhad, fso
     integer                :: numberInsertions, numHeaders, headerIndex, bodyIndex, obsNlv, obsRln
     character(len = 512)   :: queryData, queryHeader, queryCreate 
     character(len = 12 )   :: idStation
@@ -1410,15 +1412,29 @@ contains
     if ( fSQL_error( stat ) /= FSQL_OK ) write(*,*) myError//'fSQL_open: ', fSQL_errmsg( stat ),' filename: '//trim(fileName)
 
     ! Create the tables HEADER and DATA
-    queryCreate = 'create table header (id_obs integer primary key, id_stn varchar(50), lat real, lon real, &
+    if (addFSOdiag) then
+       queryCreate = 'create table header (id_obs integer primary key, id_stn varchar(50), lat real, lon real, &
+                   &codtyp integer, date integer, time integer, elev real); &
+                   &create table data (id_data integer primary key, id_obs integer, varno integer, vcoord real, &
+                   &vcoord_type integer, obsvalue real, flag integer, oma real, ompt real, oma0 real, omp real, &
+                   &an_error real, fg_error real, obs_error real, sigi real, sigo real, zhad real, fso real);'
+    else
+       queryCreate = 'create table header (id_obs integer primary key, id_stn varchar(50), lat real, lon real, &
                    &codtyp integer, date integer, time integer, elev real); &
                    &create table data (id_data integer primary key, id_obs integer, varno integer, vcoord real, &
                    &vcoord_type integer, obsvalue real, flag integer, oma real, ompt real, oma0 real, omp real, &
                    &an_error real, fg_error real, obs_error real, sigi real, sigo real, zhad real);'
+    end if
     call fSQL_do_many( db, queryCreate, stat )
     if ( fSQL_error(stat) /= FSQL_OK ) call sqlr_handleError( stat, 'fSQL_do_many with query: '//trim(queryCreate) )
 
-    queryData = 'insert into data (id_data, id_obs, varno, vcoord, vcoord_type, obsvalue, flag, oma, oma0, ompt, fg_error, obs_error, sigi, sigo, zhad) values(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);'
+    if (addFSOdiag) then
+       queryData = 'insert into data (id_data, id_obs, varno, vcoord, vcoord_type, obsvalue, flag, oma, oma0, ompt, fg_error, &
+    obs_error, sigi, sigo, zhad, fso) values(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);'
+    else
+       queryData = 'insert into data (id_data, id_obs, varno, vcoord, vcoord_type, obsvalue, flag, oma, oma0, ompt, fg_error, &
+    obs_error, sigi, sigo, zhad) values(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);'
+    end if
     queryHeader = ' insert into header (id_obs, id_stn, lat, lon, date, time, codtyp, elev ) values(?,?,?,?,?,?,?,?); '
 
     write(*,*) myName//' Insert query Data   = ', trim( queryData )
@@ -1513,6 +1529,13 @@ contains
         else
           zhad = obs_missingValue_R
         end if
+        if (addFSOdiag) then
+          if ( obs_columnActive_RB(obsdat, OBS_FSO) ) then
+            fso = obs_bodyElem_r(obsdat, OBS_FSO, bodyIndex)
+          else
+            fso = obs_missingValue_R
+          end if
+        end if
 
         select case( obsFamily )
           case ( 'UA', 'AI', 'SW' )
@@ -1579,6 +1602,13 @@ contains
         else
           call fSQL_bind_param( stmtData, PARAM_INDEX = 15, REAL_VAR = zhad )
         end if 
+        if (addFSOdiag) then
+          if ( fso == obs_missingValue_R ) then
+            call fSQL_bind_param( stmtData, PARAM_INDEX = 16                        )
+          else
+            call fSQL_bind_param( stmtData, PARAM_INDEX = 16, REAL_VAR = fso )
+          end if
+        end if
 
         call fSQL_exec_stmt ( stmtData )