+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
  * Add allsky cases for sql diagnostic files (#500 and !471)
+ * Add ability to include FSO in diagnostic sqlite files (#487 and !463)
+    * should no longer be in burp files
 
 ### Changed